pyproject.toml
[project]
name = "promenade"
version = "0.1.0"
description = "A Textual-based CLI dashboard for Prometheus metrics"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "textual>=0.47.0",
    "pyyaml>=6.0",
    "prometheus-client>=0.19.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.6.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pytest>=7.4.0",
    "types-pyyaml>=6.0.0",
    "types-requests>=2.31.0",
]

[project.scripts]
promenade = "promenade.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "Q"]
ignore = []

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict = true

[[tool.mypy.overrides]]
module = "textual.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "click.*"
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["src/promenade"]

[dependency-groups]
dev = [
    "types-click>=7.1.8",
]
